MNT: move main() around
