+++ ads_log_daemon/__main__.py
@@ -1,3 +1,3 @@
-from .client import main
+from .main import main
 
 main()
